Дипломная по Django группа DJ-5

+++ shop/views.py
@@ -2,9 +2,9 @@ from django.shortcuts import render, get_object_or_404, redirect
 from .models import Category, Article, Item, Review, Cart, ItemInCart, ItemInOrder, Order
 from .forms import AddReview
 from django.core.cache import cache
-from django.db.models.signals import post_delete, post_save
-from django.dispatch import receiver
+from .utils import *
 from django.core.paginator import Paginator
+from django.http import HttpResponseNotAllowed
 
 
 def index(request):
@@ -72,31 +72,36 @@ def add_review(request, item_id):
 
 
 def to_cart(request, item_id):
-    cart = get_cart(get_sid(request))
-    item = Item.objects.get(id=item_id)
+    if request.method == 'POST':
+        cart = get_cart(get_sid(request))
+        item = Item.objects.get(id=item_id)
 
-    item_in_cart = ItemInCart.objects.filter(cart=cart, item=item)
+        item_in_cart = ItemInCart.objects.filter(cart=cart, item=item)
 
-    if item_in_cart:
-        my_obj = item_in_cart.first()
-        my_obj.count += 1
-        my_obj.save()
+        if item_in_cart:
+            my_obj = item_in_cart.first()
+            my_obj.count += 1
+            my_obj.save()
+        else:
+            ItemInCart.objects.create(cart=cart, item=item, count=1)
     else:
-        ItemInCart.objects.create(cart=cart, item=item, count=1)
-
+        raise HttpResponseNotAllowed(['GET'])
     return redirect(cart_view)
 
 
 def create_order(request):
-    cart = get_cart(get_sid(request))
-    items = cart.items.all()
+    if request.method == 'POST':
+        cart = get_cart(get_sid(request))
+        items = cart.items.all()
 
-    if items:
-        order = Order.objects.create(owner=get_sid(request))
-        for item in items:
-            iib = ItemInCart.objects.get(item=item, cart=cart)
-            ItemInOrder.objects.create(item=item, order=order, count=iib.count)
-            iib.delete()
+        if items:
+            order = Order.objects.create(owner=get_sid(request))
+            for item in items:
+                iib = ItemInCart.objects.get(item=item, cart=cart)
+                ItemInOrder.objects.create(item=item, order=order, count=iib.count)
+                iib.delete()
+    else:
+        raise HttpResponseNotAllowed(['GET'])
     return redirect('index')
 
 
@@ -105,31 +110,8 @@ def get_cart(sid):
     return cart
 
 
-def get_sid(request):
-    if not request.user.is_authenticated:
-        sid = request.session.session_key
-        if not sid:
-            sid = request.session.cycle_key()
-    else:
-        sid = request.user.username
-    return sid
-
-
 def get_categories():
-    categories = cache.get_or_set('categories', Category.objects.filter(is_main=True), 3600)
+    categories = cache.get_or_set('categories', Category.objects.filter(is_main=True).prefetch_related(), 3600)
     return categories
 
 
-def clear_authors_count_cache():
-    cache.delete('categories')
-
-
-@receiver(post_delete, sender=Article)
-def category_post_delete_handler(sender, **kwargs):
-    clear_authors_count_cache()
-
-
-@receiver(post_save, sender=Article)
-def category_post_save_handler(sender, **kwargs):
-    if kwargs['created']:
-        clear_authors_count_cache()
